state: Move `system_call()` to separate file
